Move function to theme
