--- src/components/publiq-ui/getValueFromTheme.js
@@ -1,6 +1,0 @@
-import get from 'lodash/get';
-
-const getValueFromTheme = (component) => (path) => (props) =>
-  get(props.theme, `components.${component}.${path}`);
-
-export { getValueFromTheme };
